backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
redis==5.2.0
boto3==1.35.0
cryptography==43.0.0
pyotp==2.9.0
qrcode==8.0
reportlab==4.2.5
httpx==0.27.2
python-dateutil==2.9.0
aiosmtplib==3.0.1

